[fix] 사진 촬영, 이미지 첨부
